+++ src/main.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
 import { TestHub, testExplorerExtensionId } from 'vscode-test-adapter-api';
 import { Log, TestAdapterRegistrar } from 'vscode-test-adapter-util';
-import { ExampleAdapter } from './adapter';
+import { Z80UnitTestAdapter } from './adapter';
 
 export async function activate(context: vscode.ExtensionContext) {
 
@@ -9,7 +9,7 @@ export async function activate(context: vscode.ExtensionContext) {
 
 	// create a simple logger that can be configured with the configuration variables
 	// `exampleExplorer.logpanel` and `exampleExplorer.logfile`
-	const log = new Log('exampleExplorer', workspaceFolder, 'Example Explorer Log');
+	const log = new Log('z80UnitTestExplorer', workspaceFolder, 'Z80 Unit Test Explorer Log');
 	context.subscriptions.push(log);
 
 	// get the Test Explorer extension
@@ -23,7 +23,7 @@ export async function activate(context: vscode.ExtensionContext) {
 		// this will register an ExampleTestAdapter for each WorkspaceFolder
 		context.subscriptions.push(new TestAdapterRegistrar(
 			testHub,
-			workspaceFolder => new ExampleAdapter(workspaceFolder, log),
+			workspaceFolder => new Z80UnitTestAdapter(workspaceFolder, log),
 			log
 		));
 	}